% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_kobo_labels}
\alias{apply_kobo_labels}
\title{Apply Kobo Labels}
\usage{
apply_kobo_labels(dataset, column_name, kobo_choices)
}
\arguments{
\item{dataset}{The dataset to modify}

\item{column_name}{The name of the column to relabel}

\item{kobo_choices}{The choices dictionary from Kobo}
}
\value{
The updated dataset
}
\description{
Renames the analysis output options columns using a Kobo choices dictionary before a user can save the analysis.
}
