% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_tool_labels}
\alias{apply_tool_labels}
\title{Apply Tool Labels}
\usage{
apply_tool_labels(dataset, column_name, tool_choices)
}
\arguments{
\item{dataset}{The dataset to modify}

\item{column_name}{The name of the column to relabel}

\item{tool_choices}{The XLSForm choices sheet dataframe}
}
\value{
The updated dataset
}
\description{
Renames the analysis output options columns using an XLSForm tool choices dictionary before a user can save the analysis.
}
