% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.get_val}
\alias{.get_val}
\title{Look up current value from dataset by uuid and column}
\usage{
.get_val(
  dataset,
  uuid_column,
  uuid,
  column,
  skip_label_row = TRUE,
  uuid_index = NULL
)
}
\description{
Look up current value from dataset by uuid and column
}
\keyword{internal}
