% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.get_vocab}
\alias{.get_vocab}
\title{Get the vocabulary (all valid tokens) for a list from tool_choices}
\usage{
.get_vocab(list_name, tool_choices)
}
\description{
Get the vocabulary (all valid tokens) for a list from tool_choices
}
\keyword{internal}
