% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.remove_code}
\alias{.remove_code}
\title{Remove a code from a space-separated code string (atomic token match only)}
\usage{
.remove_code(concat, code)
}
\description{
Remove a code from a space-separated code string (atomic token match only)
}
\keyword{internal}
