% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.remove_token}
\alias{.remove_token}
\title{Remove a token from a space-joined concat using vocabulary-aware splitting}
\usage{
.remove_token(concat, token, vocab)
}
\description{
Remove a token from a space-joined concat using vocabulary-aware splitting
}
\keyword{internal}
