% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.resolve_token_to_token}
\alias{.resolve_token_to_token}
\title{Resolve a token (name or label) to its counterpart in tool_choices}
\usage{
.resolve_token_to_token(token, list_name, tool_choices, to = "label")
}
\description{
Resolve a token (name or label) to its counterpart in tool_choices
}
\keyword{internal}
