% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.sm_sub_cols}
\alias{.sm_sub_cols}
\title{Get names of select-multiple sub-columns for a parent column}
\usage{
.sm_sub_cols(dataset, parent_col, sm_separator)
}
\description{
Get names of select-multiple sub-columns for a parent column
}
\keyword{internal}
