% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.build_recode_rows}
\alias{.build_recode_rows}
\title{Build log rows for a RECODE action}
\usage{
.build_recode_rows(x, dataset, kobo_choices, uuid_column, sm_separator)
}
\description{
Build log rows for a RECODE action
}
\keyword{internal}
