% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_other_response.R
\name{.build_remove_rows}
\alias{.build_remove_rows}
\title{Build log rows for a REMOVE action}
\usage{
.build_remove_rows(
  x,
  dataset,
  uuid_column,
  sm_separator,
  skip_label_row = TRUE
)
}
\description{
Build log rows for a REMOVE action
}
\keyword{internal}
