% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_db.R
\name{get_other_db}
\alias{get_other_db}
\title{Get Other DB}
\usage{
get_other_db(kobo_survey, kobo_choices, other_labels)
}
\arguments{
\item{kobo_survey}{A dataframe representing the Kobo survey.}

\item{kobo_choices}{A dataframe representing the Kobo choices.}

\item{other_labels}{A dataframe retrieved from \code{get_other_labels}.}
}
\value{
A dataframe representing the mapping required for other responses database.
}
\description{
Processes the 'other_labels' alongside the survey inputs to map out the available choices for recoding.
}
