% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_db.R
\name{get_other_labels}
\alias{get_other_labels}
\title{Get Other Labels}
\usage{
get_other_labels(kobo_survey)
}
\arguments{
\item{kobo_survey}{A dataframe containing the Kobo survey sheet.}
}
\value{
A dataframe containing the corresponding other labels.
}
\description{
Retrieves text labels for questions in the kobo survey that correspond to "other" responses.
}
