% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_completeness.R
\name{validate_refused}
\alias{validate_refused}
\title{Validate Refused Responses}
\usage{
validate_refused(
  dataset,
  uuid_column = "uuid",
  log_name = "refused_log",
  max_refused = 6,
  refused_value = "Refused"
)
}
\arguments{
\item{dataset}{A dataframe or a list containing a dataframe named \code{checked_dataset}.}

\item{uuid_column}{The name of the column containing the unique identifier. Default is \code{"uuid"}.}

\item{log_name}{The name of the log to create in the list. Default is \code{"refused_log"}.}

\item{max_refused}{The maximum allowable number of refused responses. Default is \code{6}.}

\item{refused_value}{The response value to flag as refused. Default is \code{"Refused"}.}
}
\value{
A list containing the original dataset and the new log dataframe.
}
\description{
Checks the entire dataset and flags any records that have more than a specified
threshold of "Refused" responses.
}
